Game UI
Created the affiliated scenes and scripts for the Main Menu, Setting menu, Victory scene, Defeat scene, Pause UI and for quitting the application. Altered the PlayerHealth script so when the player's health reaches zero, they lose. Created an object which has the WinGame script and allows the player to win when they trigger with the object.